delivery_service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-jose[cryptography]
